// packages/frontend/src/generated/types/OwnershipTransferItemResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `ownership_transfer_item` entity.
 */
export interface OwnershipTransferItemResponse {
  /**
   * @type string
   */
  amount: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  fromContractorId: string
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string, uuid
   */
  productId: string
  /**
   * @type string, uuid
   */
  storageId: string
  /**
   * @type string, uuid
   */
  toContractorId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}